requirement.txt

fastapi==0.115.2
uvicorn==0.32.0
streamlit==1.39.0

torch>=2.2.0
transformers>=4.44.0
diffusers>=0.30.0
accelerate>=0.33.0

numpy>=1.26.0
Pillow>=10.0.0
requests>=2.31.0
ipywidgets>=8.1.2
gTTS>=2.5.3
nest_asyncio>=1.6.0
pyngrok>=7.1.6


